requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.14.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.2.0
gunicorn==23.0.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pipenv==2023.12.1
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-unsplash==1.2.5
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.14.1
setuptools==70.0.0
six==1.17.0
soupsieve==2.6
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.2
Werkzeug==3.1.3
xgboost==2.1.4
